+++ examples/gemm/kernels/src/gemm_naive.rs
@@ -32,15 +32,15 @@ pub unsafe fn gemm_naive(
     alpha: f32,
     beta: f32,
 ) {
-    let row = (thread::block_dim_x() * thread::block_idx_x() + thread::thread_idx_x()) as usize;
-    let col = (thread::block_dim_y() * thread::block_idx_y() + thread::thread_idx_y()) as usize;
+    let row = thread::block_dim_x() * thread::block_idx_x() + thread::thread_idx_x();
+    let col = thread::block_dim_y() * thread::block_idx_y() + thread::thread_idx_y();
 
     if row < m && col < n {
         let mut sum = 0.0f32;
         for i in 0..k {
             sum += mat_a[row * k + i] * mat_b[i * n + col];
         }
-        let elem = unsafe { &mut *mat_c.add((row * n + col) as usize) };
+        let elem = unsafe { &mut *mat_c.add(row * n + col) };
         *elem = alpha * sum + beta * *elem;
     }
 }
